+++ src/popup/App.tsx
@@ -1,21 +1,20 @@
 import { hot } from 'react-hot-loader';
 import { browser } from 'webextension-polyfill-ts';
 import React, { useState, useCallback, useEffect } from 'react';
-import { Button, FilterList, Flex, Text, ProgressBar, theme as primer, Tooltip, Flash, Link } from '@primer/components';
+import { Button, FilterList, Text, theme as primer, Flash } from '@primer/components';
 import { ThemeProvider } from 'styled-components';
-import { SyncIcon, GearIcon, CheckIcon, VersionsIcon, PeopleIcon } from '@primer/octicons-react';
+import { CheckIcon } from '@primer/octicons-react';
 import { MergeRequestItem } from './components/MergeRequestItem';
 import { IssueItem } from './components/IssueItem';
 import { TodoItem } from './components/TodoItem';
 import { EmptyItems } from './components/EmptyItems';
 import { Nav } from './components/Nav';
+import { Footer } from './components/Footer';
 import { MergeRequestsDetails, Todo, Issue } from '../background/types';
-import { getHumanReadableDate } from './helpers';
 import { getMergeRequestList, MergeRequestSendMessageReply } from './utils/mergeRequestDownloader';
+import { AppStatus } from './types';
 
 import './style.css';
-
-type AppStatus = 'idle' | 'loading' | 'success' | 'error';
 
 const App = () => {
     const [appStatus, setAppStatus] = useState<AppStatus>('idle');
@@ -33,8 +32,6 @@ const App = () => {
     const [currentTab, setCurrentTab] = useState(0);
     const [gitlabAddress, setGitlabAddress] = useState('');
     const [todosVisibility, setTodosVisibility] = useState(true);
-
-    const openSettings = useCallback(() => browser.runtime.openOptionsPage(), []);
 
     const applySettings = useCallback(() => {
         const getSettings = browser.storage.local.get(['defaultTab', 'gitlabAddress']);
@@ -144,6 +141,7 @@ const App = () => {
         if (!mrList || mrList.length === 0) {
             return <EmptyItems />;
         }
+
         return (
             <FilterList className={'mrList'}>
                 {mrList.map((mr: MergeRequestsDetails) => (
@@ -153,77 +151,18 @@ const App = () => {
         );
     }, [appStatus, mrData, currentTab, errorMessage, setAllTodosAsDone, todosVisibility]);
 
-    const getMrRatio = useCallback(() => {
-        if (appStatus !== 'success' || mrData === null) {
-            return -1;
-        }
-        let mrNumber = 0;
-        let mrTotal = 0;
-        if (currentTab === 0) {
-            mrTotal = mrData.mrAssigned.length;
-            mrNumber = mrTotal - mrData.mrToReview;
-        } else if (currentTab === 1) {
-            mrTotal = mrData.mrGiven.length;
-            mrNumber = mrData.mrReviewed;
-        } else {
-            return -1;
-        }
-        let rate = 0;
-        if (mrNumber && mrTotal) {
-            rate = mrNumber / mrTotal;
-        }
-        return Math.floor(rate * 100);
-    }, [appStatus, mrData, currentTab]);
-
     return (
         <ThemeProvider theme={primer}>
             <div className={'container'}>
                 <Nav currentTab={currentTab} setCurrentTab={setCurrentTab} mrData={mrData} />
                 {getContent()}
-                <Flex flexWrap="nowrap">
-                    <Tooltip
-                        className={'progressBar'}
-                        aria-label={
-                            getMrRatio() >= 0
-                                ? `${getMrRatio()}% reviewed, keep the good work!`
-                                : 'Nothing to measure here...'
-                        }
-                        direction="n"
-                    >
-                        <ProgressBar progress={getMrRatio()} />
-                    </Tooltip>
-
-                    <div style={{ marginTop: '6px' }}>
-                        <Tooltip
-                            aria-label={'Last update: ' + getHumanReadableDate(mrData.lastUpdateDateUnix)}
-                            direction="n"
-                        >
-                            <Button onClick={fetchData} variant={'small'} mr={2}>
-                                <SyncIcon /> Refresh
-                            </Button>
-                        </Tooltip>
-
-                        <Tooltip aria-label={'Open your projects'} direction="n">
-                            <Link href={gitlabAddress + '/dashboard/projects'} target="_blank">
-                                <Button variant={'small'} mr={2}>
-                                    <VersionsIcon />
-                                </Button>
-                            </Link>
-                        </Tooltip>
-
-                        <Tooltip aria-label={'Open your groups'} direction="n">
-                            <Link href={gitlabAddress + '/dashboard/groups'} target="_blank">
-                                <Button variant={'small'} mr={2}>
-                                    <PeopleIcon />
-                                </Button>
-                            </Link>
-                        </Tooltip>
-
-                        <Button onClick={openSettings} variant={'small'}>
-                            <GearIcon /> Options
-                        </Button>
-                    </div>
-                </Flex>
+                <Footer
+                    currentTab={currentTab}
+                    mrData={mrData}
+                    appStatus={appStatus}
+                    fetchData={fetchData}
+                    gitlabAddress={gitlabAddress}
+                />
             </div>
         </ThemeProvider>
     );
