+++ src/popup/types.ts
@@ -1,0 +1,1 @@
+export type AppStatus = 'idle' | 'loading' | 'success' | 'error';
